litfusion_api.py
from PIL import Image
from litserve import LitServer, LitAPI

from openai_image_spec import OpenAIImageSpec


class LitOpenAIImage(LitAPI):
        def setup(self, device):
            # Placeholder setup for loading image generation models
            pass

        def predict(self, request):
            # Logic to determine which type of request it is
            request_type = request.get('request_type')
            if request_type == "generation":
                yield Image.new("RGB", (256, 256), color="blue")  # Example generated image
            elif request_type == "edit":
                yield Image.new("RGB", (256, 256), color="green")  # Example edited image
            elif request_type == "variation":
                yield Image.new("RGB", (256, 256), color="red")  # Example variation image
            else:
                yield "Unknown request type"

api = LitOpenAIImage()
server = LitServer(api, spec=OpenAIImageSpec())
server.run(port=8000)