requirements.txt
litserve
pillow